Move cmdmgr to 3 different files like other managers (eg: servstate)
