+++ internal/overlord/cmdstate/manager.go
@@ -1,0 +1,32 @@
+// Copyright (c) 2021 Canonical Ltd
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License version 3 as
+// published by the Free Software Foundation.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+package cmdstate
+
+import (
+	"github.com/canonical/pebble/internal/overlord/state"
+)
+
+type CommandManager struct{}
+
+// NewManager creates a new CommandManager.
+func NewManager(runner *state.TaskRunner) *CommandManager {
+	runner.AddHandler("exec", doExec, nil)
+	return &CommandManager{}
+}
+
+// Ensure is part of the overlord.StateManager interface.
+func (m *CommandManager) Ensure() error {
+	return nil
+}
